refactor: add KeywordArgument interface to replace []string in keywordArgument map.
this segments multi-arg keywords when they are used multiple times. for example if you use the foreignkey keyword multiple times on a single field, the arguments would've been in one contiguous slice []string. using a slice of []KeywordArguments we know that both arguments for foreignkey are in KeywordArguments[n]. this also enables multi-arg validation checks.

- splits keywords into their own file
- collapses two test functions into a single function